+++ dvc/config.py
@@ -17,7 +17,7 @@ from .utils.objects import cached_property
 
 if TYPE_CHECKING:
     from dvc.fs import FileSystem
-    from dvc.types import DictStrAny
+    from dvc._types import DictStrAny
 
 logger = logger.getChild(__name__)
