+++ dvc/repo/worktree.py
@@ -14,7 +14,7 @@ if TYPE_CHECKING:
     from dvc.repo.index import Index, IndexView
     from dvc.repo.stage import StageInfo
     from dvc.stage import Stage
-    from dvc.types import TargetType
+    from dvc._types import TargetType
     from dvc_data.hashfile.meta import Meta
     from dvc_data.index import DataIndex, DataIndexView
     from dvc_objects.fs.base import FileSystem
