+++ dvc/stage/__init__.py
@@ -36,7 +36,7 @@ if TYPE_CHECKING:
     from dvc.dvcfile import ProjectFile, SingleStageFile
     from dvc.output import Output
     from dvc.repo import Repo
-    from dvc.types import StrPath
+    from dvc._types import StrPath
     from dvc_data.hashfile.db import HashFileDB
     from dvc_data.hashfile.hash_info import HashInfo
     from dvc_objects.db import ObjectDB
