+++ dvc/testing/benchmarks/fixtures.py
@@ -11,7 +11,7 @@ from dulwich.porcelain import clone
 from funcy import first
 from packaging import specifiers, version
 
-from dvc.types import StrPath
+from dvc._types import StrPath
 
 
 @pytest.fixture(scope="session")
