+++ dvc/utils/serialize/_common.py
@@ -10,7 +10,7 @@ from dvc.exceptions import DvcException
 
 if TYPE_CHECKING:
     from dvc.fs import FileSystem
-    from dvc.types import StrPath
+    from dvc._types import StrPath
 
 
 class DumperFn(Protocol):
